ref: extracted Agent response container
